Update _source function to source files recursively
The motivation for this is if I want to add custom bash completions,
I can create a shell_sources/completions directory and put files there
and have them automatically sourced without changing any other code.

Since I used a recursive glob, which is Bash 4 only, I removed some
of the guards on other Bash 4 features. Bash 4 was released in 2009,
it should be safe to depend on at this point.