Change Sublime Text theme back to defaults
Transferring Sublime Text settings over to a new computer was
un-reproducible, because the theme filename is changed by
extensions as they add things. This commit switches Sublime Text
theme and color scheme settings back to defaults so that it'll
work the same everywhere.

I've switched to VSCode which has an extension to sync settings, so:
* Sublime Text settings now matter less
* I won't have this problem again