Fix 'lightning bolt' size in prompt (iterm 3.1.6)
iTerm2 3.1.6 "fixed" Unicode width ("Use Unicode 9 widths by default")
that caused the prompt to bug out because iTerm2 was confused about
how wide the character was. This was a related problem to what was
happening on EC2 that was addressed in 364ed41, but now it's fixed
the right way everywhere.

Turns out Terminal.app was doing it the right way all along, which
is why the lightning bolt always looked different on it (I'd needed
to pad it with a space afterwards for iTerm2 but not for Terminal).