Make 'uc' and 'lc' scripts instead of aliases
This enables use for case-insensitive comparisons, with map, etc.

The 'set -- ...' trick turns anything provided on stdin into
arguments if no arguments are passed. This lets uc/lc be called
in a pipeline or with arguments.
https://stackoverflow.com/questions/19619490/how-to-read-stdin-when-no-arguments-are-passed/31602858#comment78199557_31602858

I used awk vs tr, sed, etc. because it seems to be the most
portable and Unicode-compatible: https://askubuntu.com/a/383819