Fix suspended job count in prompt when working directory changes
In zsh, when you've changed your working directory, jobs looks like:

$ jobs | head
[1]    suspended  vim
(pwd now: ~/proj)
[2]    running    sleep 100000
[3]  - suspended  vim
[4]  + suspended  vim

So the extra line "(pwd now:...)" caused the line count to be off when
the working directory changed. However, this is only when piping 'jobs'
to something, for some reason! If you run 'jobs' directly it doesn't
show the "(pwd..."

In bash, jobs looks like:

$ jobs
[1]-  Stopped                 vim  (wd: ~)
[2]   Running                 sleep 100000 &  (wd: ~)
[3]+  Stopped                 vim

So the previous code was always correct for bash.