Strip down bash config; make it easier to run remotely
Bash will never be a main shell again, so strip down the config to
lowest common denominator, and gate things like git aliases behind an
"is local" check to make it easier to run over ssh or in a Docker
container.

Also include a preliminary 'docker-bash' script to start up a Docker
container with your bash config. Skip locale init in Perl call in prompt
since it's not necessary for that call and fails if locales aren't set
up on container/server.