Poetry: create virtualenvs in-project
Reason: currently having problems using Syncthing to work on the
same project from multiple computers because the virtualenvs are
created at the system level and are named differently.

eg. vscode gets confused about the Python installation because
the name of the virtualenv differs between computers.

This way, projects should be able to share virtualenvs via Syncthing.