kw: add -d/--dir to run command in window at 'dir'
Search for an existing window at 'dir' that's open to your shell,
or if there is none, open a new tab to launch the command in.

Also:
- add special case to output the kitty socket path
- add -t to open in new tab (exclusive with -H/v/n)

todo: the code can be improved, eg. when opening in --dir, it currently
ignores the existing location arguments and always opens in a new tab.