Prepare for M2 Mac; account for different Homebrew paths
TIL `brew --prefix` is fast now thanks to being bash instead of Ruby,
so no problem calling it in startup scripts anymore.

Because shebangs have to be an absolute path, hard-coding the path to
Homebrew'd Python won't work anymore. Add a symlink to the Homebrew python
to your ~/bin to go back to using /usr/bin/env in shebangs.

I'm not sure the solution to this in general for Python scripts.