aliases: create daily note alias
Yes the notes/notes is correct, that's an artifact of how Dendron
creates its daily notes, stick with it while we're still using Dendron.

It'll be fun when we figure out what the best bulk rename cli tool is
when we pick a new name format.

todo: see if Dendron adds its metadata when it opens the file if you
create a new daily note for it in this fashion.

question: should you call this "diary" in honor of Andy's "Dear Diary"?