git: remove notes.txt from global git ignore
You'd typically create a notes.txt as one of the first actions in a
new repository. While convenient, this leads to a situation where notes
could be spread between your centralized system and different projects.
This makes them harder to search, and so on. Worse, those notes wouldn't
be source controlled or backed up and so could be lost or forgotten easily.

After removing it from the global ignore, I discovered this repository
even had a trivial notes.txt that I didn't realize I had sitting around.

Stop this practice. From now on keep all project notes in centralized notes.