vscode: binrun.show should work if the shell is focused
You have it prominently bound currently to F3.
With the terminal focused, pressing F3 would input 'R' (??)
With this setting, it passes F3 through.

https://code.visualstudio.com/docs/terminal/advanced#_keybinding-and-the-shell

I was surprised that this was how to accomplish this. I'd have expected
to remove the keyboard shortcut for the terminal F3, or set a 'when'
that gives the user global one higher priority, idk.