vscode: bind ctrl+1 to focus open files pane
ctrl+1 through ctrl+9 are based on tab order, which is pointless and
arbitrary now that I've hidden tabs. I would like them to go based on
the order in my open files pane but there's no built in way to do that.

I feel like something should be possible with 'runCommands', focusing
the pane, then doing the equivalent of sending HOME, down*N, enter
keystrokes to open the nth editor in that pane.

I couldn't get that to work yet so for now just bind ctrl+1 to focus the
open editors pane and you can navigate manually from there.