zsh: set insert-tab false
This means that when you're sitting at a blank terminal,
it still calls autocomplete instead of inserting a literal
tab character, which is never what you want.

It also means that when you're writing a multi-line function
interactively 🙄 and hit tab to indent, instead of a literal
tab you get autocomplete, which is also not what you want.

Overall, this is an improvement from "insert literal tab",
so leaving on, but look for a more useful "tab at empty prompt"
behavior. That's prime UI "real estate" waiting to be used.