Turn 'setup' into a just script
This brings 'setup' full circle. It is now a script of its own again,
just in a different language (just).

It makes sense to be its own script vs either the project's justfile
(since it's not really for the setup project itself) or the system-level
justfile (since that makes me uncomfortable).