python: uv venv --allow-existing
uv started issuing a warning about an existing venv.
Non-interactive options are to --allow-existing or --clear.

In the manpage there's a warning:

> WARNING: This option can lead to unexpected behavior if the existing
> virtual environment and the newly-created virtual environment are
> linked to different Python interpreters.

So, between Python versions I may have a problem?

I'm content to wait until the next major Python version change to see if
it causes a problem. I'd expect uv to call out if it was installing into
a venv of a different major Python version?

Maybe it's better to clear the venv each time?